add stop drawing button, toggle button states
Added a stop drawing button to escape current drawing command without erasing previously created graphics.  Also disabled the clearing of graphics on widget close - felt like this was unexpected by users and there is a button to do that when needed.  Finally, added toggling of the stop and reset drawing buttons - perhaps overkill but wanted to test the user interaction.